requirements.txt
groq
python-dotenv
langchain-groq
fastapi
pdfplumber
langdetect
deep-translator
uvicorn
python-multipart